moved error component to routes
