--- packages/cashier/src/components/cashier-error/index.js
@@ -1,3 +1,0 @@
-import CashierError from './cashier-error.jsx';
-
-export default CashierError;
