+++ packages/cashier/src/components/error/error.scss
