+++ packages/cashier/src/components/error/index.js
@@ -1,0 +1,3 @@
+import Error from './error.jsx';
+
+export default Error;
