+++ packages/cashier/src/containers/routes/error-component/__tests__/error-component.spec.js
