+++ packages/cashier/src/containers/routes/routes.jsx
@@ -4,7 +4,7 @@ import React from 'react';
 import { withRouter } from 'react-router';
 import { connect } from 'Stores/connect';
 import BinaryRoutes from './binary-routes.jsx';
-import ErrorComponent from 'Components/error-component';
+import ErrorComponent from './error-component';
 
 const Routes = ({ error, has_error, is_logged_in, is_logging_in, passthrough }) => {
     if (has_error) {
