+++ packages/cashier/src/pages/payment-agent-transfer/__tests__/payment-agent-transfer.spec.js
@@ -20,7 +20,7 @@ jest.mock('@deriv/components', () => {
 });
 
 jest.mock('Components/cashier-locked', () => jest.fn(() => 'mockedCashierLocked'));
-jest.mock('Components/cashier-error', () => jest.fn(() => 'mockedCashierError'));
+jest.mock('Components/error', () => jest.fn(() => 'mockedError'));
 jest.mock('Components/no-balance', () => jest.fn(() => 'mockedNoBalance'));
 jest.mock('Pages/payment-agent-transfer/payment-agent-transfer-form', () =>
     jest.fn(() => 'mockedPaymentAgentTransferForm')
@@ -89,7 +89,7 @@ describe('<PaymentAgentTransfer />', () => {
 
         render(<PaymentAgentTransfer {...props} error={cta_error} />);
 
-        expect(screen.getByText('mockedCashierError')).toBeInTheDocument();
+        expect(screen.getByText('mockedError')).toBeInTheDocument();
     });
 
     it('should show the no balance component if account has no balance', () => {
