+++ packages/cashier/src/pages/payment-agent/payment-agent-withdrawal-locked/payment-agent-withdrawal-locked.jsx
@@ -4,7 +4,7 @@ import { Button, StaticUrl, Text } from '@deriv/components';
 import { localize, Localize } from '@deriv/translations';
 import { routes } from '@deriv/shared';
 import { withRouter } from 'react-router-dom';
-import CashierError from 'Components/cashier-error';
+import Error from 'Components/error';
 import './payment-agent-withdrawal-locked.scss';
 
 const PaymentAgentWithdrawalLockedItem = ({ item }) => {
@@ -62,7 +62,7 @@ const PaymentAgentWithdrawalLocked = ({ error, history }) => {
     ];
 
     if (error.onClickButton || error.code === 'PaymentAgentJustification') {
-        return <CashierError error={error} />;
+        return <Error error={error} />;
     }
 
     return (
